requirements.txt
langchain-text-splitters
langchain-core
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
pymupdf
python-dotenv
requests
pydantic
